test: Spliting default tests in different files and also improving 'default' value tests
